Add confirmation dialog for coupon code deletion using SweetAlert2

+++ src/page/CouponCode/CouponCode.jsx
@@ -3,6 +3,7 @@ import { Modal, Form, Input, Select, DatePicker, Pagination, message } from "ant
 import dayjs from "dayjs";
 import moment from "moment";
 import { useCreateCouponCodeMutation, useDeleteCouponCodeMutation, useEditCouponCodeMutation, useGenerateCouponCodeMutation, useGetCouponCodeQuery } from "../../redux/features/couponCode/couponCode";
+import Swal from "sweetalert2";
 
 export default function CouponCode() {
     const page = 1;
@@ -107,17 +108,34 @@ export default function CouponCode() {
 
     const handleDelete = async (item) => {
         console.log(item)
-        try {
-            const response = await deleteCouponCode(item?.referralCode).unwrap();
-            console.log(response)
-            if (response) {
-                message.success("Delete Successfully !!");
-                refetch();
+
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                try {
+                    const response = await deleteCouponCode(item?.referralCode).unwrap();
+                    console.log(response)
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
+                    });
+                    refetch();
+                } catch (error) {
+                    console.error(error);
+                    message.error(error?.data?.message);
+                }
             }
-        } catch (error) {
-            console.error(error);
-            message.error(error?.data?.message);
-        }
+        });
+
+
     };
     const [generatedItem, setGeneratedItem] = useState(null)
     const [genCodeName, setGenCodeName] = useState("")
